requirements.txt
Django>=5.1
whitenoise>=6.8
pillow>=11.0
python-decouple>=3.8
psycopg2-binary>=2.9
gunicorn>=23.0
daphne>=4.1
channels>=4.1
channels-redis>=4.2
django-axes>=7.0
django-csp>=3.8
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
requests>=2.32
